backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
django-filter==23.3
gunicorn==21.2.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.0